+++ src/ui/scripts/components/cards/CardFacts.js
@@ -6,7 +6,9 @@ import formatNumber from '../../utils/formatNumber'
 import pluralize from '../../utils/pluralize'
 
 import Headline from '../Headline'
-import PresentationValueText from '../presentations/PresentationValueText'
+import TextBadge from '../TextBadge'
+import ChangeBadge from '../ChangeBadge'
+import PresentationValueUnit from '../presentations/PresentationValueUnit'
 
 const Presentation = (props) => {
 	return (
@@ -14,11 +16,19 @@ const Presentation = (props) => {
 			h(Headline, {
 				type: 'h2',
 				size: 'small',
+				className: 'facts__top',
 			}, props.headline),
-			h(PresentationValueText, {
-				value: props.value,
-				text: props.text,
-			}),
+			h('div', {
+				className: 'facts__left',
+			},
+				h(PresentationValueUnit, {
+					value: props.value,
+					unit: props.unit,
+				}),
+			),
+			props.addition != null && h('div', {
+				className: 'facts__right',
+			}, props.addition),
 		)
 	)
 }
@@ -29,7 +39,9 @@ const CardFacts = (props) => {
 	const {
 		activeVisitors,
 		averageViews,
+		averageViewsChange,
 		averageDuration,
+		averageDurationChange,
 		viewsToday,
 		viewsMonth,
 		viewsYear,
@@ -42,32 +54,35 @@ const CardFacts = (props) => {
 			h(Presentation, {
 				headline: 'Active visitors',
 				value: activeVisitors,
-				text: pluralize([ 'visitors', 'visitor', 'visitors' ], activeVisitors),
+				unit: pluralize([ 'visitors', 'visitor', 'visitors' ], activeVisitors),
+				addition: h(TextBadge, { type: 'positive', value: 'Live' }),
 			}),
 			h(Presentation, {
 				headline: 'Average views',
 				value: formatNumber(averageViews),
-				text: 'per day',
+				unit: 'per day',
+				addition: averageViewsChange != null && h(ChangeBadge, { value: averageViewsChange }),
 			}),
 			h(Presentation, {
 				headline: 'Average duration',
 				value: formatDuration(averageDuration).value,
-				text: formatDuration(averageDuration).unit,
+				unit: formatDuration(averageDuration).unit,
+				addition: averageDurationChange != null && h(ChangeBadge, { value: averageDurationChange }),
 			}),
 			h(Presentation, {
 				headline: 'Views today',
 				value: formatNumber(viewsToday),
-				text: pluralize([ 'views', 'view', 'views' ], viewsToday),
+				unit: pluralize([ 'views', 'view', 'views' ], viewsToday),
 			}),
 			h(Presentation, {
 				headline: 'Views this month',
 				value: formatNumber(viewsMonth),
-				text: pluralize([ 'views', 'view', 'views' ], viewsMonth),
+				unit: pluralize([ 'views', 'view', 'views' ], viewsMonth),
 			}),
 			h(Presentation, {
 				headline: 'Views this year',
 				value: formatNumber(viewsYear),
-				text: pluralize([ 'views', 'view', 'views' ], viewsYear),
+				unit: pluralize([ 'views', 'view', 'views' ], viewsYear),
 			}),
 		)
 	)