+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This release contains a refactored front-end that takes advantage of the GraphQL
 ### Added
 
 - Views and duration details: Click on a chart bar on the overview and insights page to see more details
+- Percentage changes of average views and duration in the facts panel
 - Active visitors counter updates periodically without reloading the UI
 - Tooltips for text in lists (#266)